Implement proper cc / bcc handling.
The BCC part was tricky, mainly due to the limitations of Net::SMTP.

In order to properly send an email with a BCC, you had to avoid
adding a header altogether, otherwise it won't be a BCC at all :-)

The solution was to subclass MailFactory and implement a clean
and invisible method that doesn't auto-add to headers.

Then emailing via SMTP will just add that bcc email.